Connection selectors don't need their own package.
