+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/connections/RoundRobinConnectionSelector.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package uk.ac.manchester.spinnaker.connections.selectors;
+package uk.ac.manchester.spinnaker.connections;
 
 import static java.util.Collections.unmodifiableList;
