+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/ApplicationRunProcess.java
@@ -18,8 +18,8 @@ package uk.ac.manchester.spinnaker.transceiver;
 
 import java.io.IOException;
 
+import uk.ac.manchester.spinnaker.connections.ConnectionSelector;
 import uk.ac.manchester.spinnaker.connections.SCPConnection;
-import uk.ac.manchester.spinnaker.connections.selectors.ConnectionSelector;
 import uk.ac.manchester.spinnaker.machine.ChipLocation;
 import uk.ac.manchester.spinnaker.machine.CoreSubsets;
 import uk.ac.manchester.spinnaker.messages.model.AppID;
