+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/BasicSCPCommandProcess.java
@@ -21,8 +21,8 @@ import static uk.ac.manchester.spinnaker.connections.SCPRequestPipeline.SCP_TIME
 
 import java.io.IOException;
 
+import uk.ac.manchester.spinnaker.connections.ConnectionSelector;
 import uk.ac.manchester.spinnaker.connections.SCPConnection;
-import uk.ac.manchester.spinnaker.connections.selectors.ConnectionSelector;
 import uk.ac.manchester.spinnaker.messages.scp.CheckOKResponse;
 import uk.ac.manchester.spinnaker.messages.scp.SCPRequest;
