+++ SpiNNaker-comms/src/main/java/uk/ac/manchester/spinnaker/transceiver/FixedRouteControlProcess.java
@@ -18,8 +18,8 @@ package uk.ac.manchester.spinnaker.transceiver;
 
 import java.io.IOException;
 
+import uk.ac.manchester.spinnaker.connections.ConnectionSelector;
 import uk.ac.manchester.spinnaker.connections.SCPConnection;
-import uk.ac.manchester.spinnaker.connections.selectors.ConnectionSelector;
 import uk.ac.manchester.spinnaker.machine.HasChipLocation;
 import uk.ac.manchester.spinnaker.machine.RoutingEntry;
 import uk.ac.manchester.spinnaker.messages.model.AppID;
